func/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
persephone-client==1.3.2
heimdall-client==2.2.0
etria-logger==0.1.9

motor==2.5.1
flask==2.0.3
pydantic==1.9.0
python-decouple==3.6
aiohttp==3.8.1
aioredis==2.0.1
cx_Oracle==8.3.0
cx_Oracle_async==0.3.3
nest_asyncio==1.5.4
